requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
comtypes==1.1.11
discord-py==2.5.2
frozenlist==1.6.0
idna==3.10
MouseInfo==0.1.3
multidict==6.4.4
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==9.0.0
propcache==0.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pygrabber==0.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==1.1.0
pytweening==1.0.4
pywin32==303
requests==2.32.3
typing-extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
flask
pytz